browser/base/content/test/test_bug364677.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=364677
-->
<head>
  <title>Test for Bug 364677</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=364677">Mozilla Bug 364677</a>
<p id="display"><iframe id="testFrame" src="bug364677-data.xml"></iframe></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 364677 **/
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  // Need privs because the feed seems to have an about:feeds principal or some
  // such.  It's not same-origin with us in any case.
  netscape.security.PrivilegeManager.enablePrivilege("UniversalBrowserRead");
  is($("testFrame").contentDocument.documentElement.id, "feedHandler",
     "Feed served as text/xml without a channel/link should have been sniffed");
});
addLoadEvent(SimpleTest.finish);
</script>
</pre>
</body>
</html>
